+++ init.c
@@ -511,7 +511,8 @@ static void dump_options(metafile_t *m)
                 #endif
 
                 "  Be verbose:    yes\n",
-                m->torrent_name, m->metainfo_file_path, m->piece_length
+                m->torrent_name, realpath(m->metainfo_file_path, NULL),
+                m->piece_length
 
                 #ifdef USE_PTHREADS
                 ,m->threads